Py-SMS/Backend/requirements.txt
# Django Core
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Environment
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
